% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urltools.R
\name{as_absolute}
\alias{as_absolute}
\title{transform url to absolute}
\usage{
as_absolute(url, base)
}
\arguments{
\item{url}{character. must be valid, preferably href or src, see \verb{relative <-} in function body}

\item{base}{string. base url for all relative url, must contain protocol}
}
\value{
character. Invalid values are removed and the original order isn't
granted
}
\description{
transform url to absolute
}
