% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as_url2file}
\alias{as_url2file}
\title{transfrom a url to link to a file}
\usage{
as_url2file(url)
}
\arguments{
\item{url}{string. mustn't contain tailing '/', because in this situation there is no need to distinguish www.foo.com/bar/index.html from www.foo.com/bar (www.foo.com/bar/ must be www.foo.com/bar/index.html)}
}
\value{
string. \code{paste0(url, '/')} if it's accessible (there exist a \code{index.html} file)
}
\description{
This is the \strong{core} function, and why I want to write this
package. wget just download www.foo.com/bar/index.html to ./www.foo.com/bar
, and then overwrite that file to a folder while downloading other HTML
files in www.foo.com/bar/ . I just want to ask whether you want to cry.
}
\section{example}{


as_url2file('www.cookbook-r.com/Graphs');
as_url2file('www.cookbook-r.com/Graphs2'); # not exist, for testing
}

