% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urltools.R
\name{base_name}
\alias{base_name}
\title{get base directory name of url}
\usage{
base_name(url, tailing = T)
}
\arguments{
\item{url}{character.}
}
\value{
character. with tailing '/'
}
\description{
\code{base_name()} get base directory name of url. In other words,
extracts protocal (http or https) and domain name from url it It works
similar to but a little different from \link[base:basename]{basename()}.
}
\details{
Although www.foo.com/bar must actually be www.foo.com/bar/index.html, but it's also possible that \code{bar} is a file under domain root directory. So \code{base_name()} has to return www.foo.com
}
