% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urltools.R
\name{belong_to}
\alias{belong_to}
\title{whether the url belong to a domain}
\usage{
belong_to(url, domain)
}
\arguments{
\item{url}{character.}

\item{domain}{string. must include protocol if url does, with tailing '/'.}
}
\value{
logical
}
\description{
whether the url belong to a domain (including protocol)
}
