% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dir_create}
\alias{dir_create}
\title{recursive create directory it not exist}
\usage{
dir_create(dir)
}
\arguments{
\item{dir}{string.}
}
\value{
logical scalar. invisible
}
\description{
recursive create directory it not exist
}
\section{example}{


dir_create('data-raw/test');
}

