% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urltools.R
\name{dir_name}
\alias{dir_name}
\title{get directory name of url or path}
\usage{
dir_name(url)
}
\arguments{
\item{path}{character.}
}
\value{
character. with tailing '/'
}
\description{
\code{dir_name()} get directory name of url or path. It works similar to but a little different from \link[base:dirname]{dirname()}.
}
\details{
It doesn't works on single file name, i.e. \code{dir_name('foo/bar')} is okay but \code{dir_name('foobar')} doesn't return what you want. Since it is used to transform relative url to absolute url, \code{dir_name('www.github.com')} must return \code{'www.github.com'}.
}
