% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rget.R
\name{download_dependencies}
\alias{download_dependencies}
\title{download dependencies for a website}
\usage{
download_dependencies(website, ...)
}
\arguments{
\item{website}{\code{\link[=rget]{rget()}}}
}
\description{
download dependencies for a website after all HTML files are downloaded
}
