% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rget.R
\name{.queue}
\alias{.queue}
\title{wanted HTML file (\strong{absolute} url) queue}
\format{A list of three character
\enumerate{
\item current files to be downloaded
\item succeed files successfully downloaded
\item failed files failed to be downloaded
}}
\usage{
.queue()
}
\description{
wanted HTML file (\strong{absolute} url) queue
}
\details{
the url can contain http:// or not.

But it must be HTML files, and actually linked by other files (? and # removed). That is to
say,
\itemize{
\item if some file links to www.foo.com/bar, you shouldn't add
www.foo.com/bar/index.html to the queue
\item if some file links to /bar.html, you shouldn't add www.foo.com/bar.html
to the queue
}

Thus ensure \code{setdiff(queue$current, queue$succeed)} makes sense
}
