% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urltools.R
\name{remove_protocol}
\alias{remove_protocol}
\title{remove 'protocol://' in url}
\usage{
remove_protocol(url)
}
\arguments{
\item{url}{character.}
}
\value{
character.
}
\description{
remove 'protocol://' in url
}
\note{
https://tools.ietf.org/html/rfc3986#section-2.2

see \link{https://tools.ietf.org/html/rfc3986#appendix-B} for URL specify
}
