% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rget.R
\name{rget}
\alias{rget}
\title{download a static website}
\usage{
rget(website, dest_dir, force = FALSE, verbose = TRUE)
}
\arguments{
\item{website}{string. url of the website to be cloned, it should be the domain name (sub directory not supported yet). you can't append ?...#..., for
example, 'www.foobar.com?haha#xixi' is forbidden,}

\item{dest_dir}{string. No tailling '/'}

\item{force}{\code{\link[=download_file]{download_file()}}}

\item{verbose}{\code{\link[=download_file]{download_file()}}}
}
\description{
download a static website
}
\details{
we assume:
\enumerate{
\item static http(s) website, no query (?foo=bar), no account authentication
\item HTML file is always linked by <a> tag
\item HTML file name all end with '.html', and default file name is
'index.html'
}

rget('www.cookbook-r.com', 'data-raw'); rget('www.cookbook-r.com',
'~/Git/mirrors')
}
\section{motivation}{
 \code{\link[=as_url2file]{as_url2file()}}
}

