% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rget.R
\name{rget_html}
\alias{rget_html}
\title{process a html file a time}
\usage{
rget_html(queue, ...)
}
\arguments{
\item{queue}{\link[=.queue]{queue}}

\item{...}{other arguments passed on to \code{\link[=download_file]{download_file()}}}
}
\value{
queue
}
\description{
process a html file a time
}
