% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urltools.R
\name{shorten_url}
\alias{shorten_url}
\title{remove '?...#...' from url}
\usage{
shorten_url(url)
}
\arguments{
\item{url}{character.}
}
\value{
character.
}
\description{
remove query (?...) and fragment (#...) from url
}
